[joy-06] quick 백준 문제 풀기
